:recycle: move services and repositories for their src folders
