+++ backend/src/data/DTO/gift.ts
@@ -1,0 +1,3 @@
+import { Gift } from "../../domain/entities";
+
+export type GiftDTO = Gift;
