+++ backend/src/data/DTO/index.ts
@@ -1,0 +1,1 @@
+export * from "./gift";
