+++ backend/src/data/contracts/gift/index.ts
@@ -1,0 +1,1 @@
+export * from "./save-gift-repository";
