--- backend/src/domain/repositories/gift/index.ts
@@ -1,1 +1,0 @@
-export * from "./save-gift";
