+++ backend/src/domain/useCases/gift/index.ts
@@ -1,0 +1,1 @@
+export * from "./create-gift";
