+++ backend/tests/domain/repositories/gift/save-gift.ts
@@ -1,5 +1,5 @@
 import { Gift } from ".././../../../src/domain/entities";
-import { SaveGiftRepository } from ".././../../../src/domain/repositories/gift";
+import { SaveGiftRepository } from ".././../../../src/data/contracts/gift";
 
 export class SaveGiftRepositoryMock implements SaveGiftRepository {
   input: Gift;
